chore: move event writer logic to new file
